refactor(frontdesk): split FleetSyncWizard into fleetSync/ modules
FleetSyncWizard.tsx carried the wizard, its resting screen and every step
and badge component in one 1238-line file. It now mirrors the alerts
wizard's shape under components/fleetSync/:

- FleetSyncWizard.tsx: the state machine and the confirm modals (544)
- steps.tsx: the three step screens, TokenField, ConfigLegend,
  MemberTable, MemberBadge and WizardNav (516)
- resting.tsx: StepResting and UsageSection (201)
- gates.ts: the Step type, STEPS, and the shared fleet-status predicates
  (reachablePeers, masterKeyBlockers, schemaBlockers, configChanges,
  offlinePeers) so the wizard and the steps read the same gates (32)

Pure moves, no behaviour change. The test file moves alongside; the
suite still exercises every piece through the wizard. The size
allowlist drops the FleetSyncWizard.tsx entry since nothing is over the
ceiling any more.
